[cleanup]: 'size_t' -> 'uint64_t' fixed size ints

+++ include/mem/alloc.h
@@ -1,7 +1,7 @@
-#include <stddef.h>
+#include <stdint.h>
 
-void *kmalloc(size_t size);
-void *krealloc(void *ptr, size_t size);
-void *kzalloc(size_t size);
+void *kmalloc(uint64_t size);
+void *krealloc(void *ptr, uint64_t size);
+void *kzalloc(uint64_t size);
 void kfree(void *ptr);
 #pragma once
