+++ include/mem/bitmap_alloc.h
@@ -2,9 +2,9 @@
 #include <stddef.h>
 #include <stdint.h>
 
-void bitmap_bitmap_init(uintptr_t base_address, size_t total_pages);
-void *bitmap_alloc_pages(size_t count);
-void bitmap_free_pages(void *address, size_t count);
+void bitmap_bitmap_init(uintptr_t base_address, uint64_t total_pages);
+void *bitmap_alloc_pages(uint64_t count);
+void bitmap_free_pages(void *address, uint64_t count);
 void set_map_location();
 void unset_map_location();
 void bitmap_print_memory_status();
@@ -18,8 +18,8 @@ void bitmap_print_memory_status();
 struct vmalloc_bitmap {
     uint64_t *bitmap;
     uintptr_t base_address;
-    size_t total_pages;
-    size_t free_pages;
+    uint64_t total_pages;
+    uint64_t free_pages;
 };
 
 #pragma once
