+++ include/misc/dbg.h
@@ -4,5 +4,5 @@ void debug_print_registers();
 
 void debug_print_stack();
 
-void debug_print_memory(void *addr, size_t size);
+void debug_print_memory(void *addr, uint64_t size);
 #pragma once
