+++ kernel/misc/string.c
@@ -2,37 +2,37 @@
 #include <stddef.h>
 #include <stdint.h>
 
-void *memcpy(void *dest, const void *src, size_t n) {
+void *memcpy(void *dest, const void *src, uint64_t n) {
     uint8_t *pdest = (uint8_t *) dest;
     const uint8_t *psrc = (const uint8_t *) src;
 
-    for (size_t i = 0; i < n; i++) {
+    for (uint64_t i = 0; i < n; i++) {
         pdest[i] = psrc[i];
     }
 
     return dest;
 }
 
-void *memset(void *s, int c, size_t n) {
+void *memset(void *s, int c, uint64_t n) {
     uint8_t *p = (uint8_t *) s;
 
-    for (size_t i = 0; i < n; i++) {
+    for (uint64_t i = 0; i < n; i++) {
         p[i] = (uint8_t) c;
     }
 
     return s;
 }
 
-void *memmove(void *dest, const void *src, size_t n) {
+void *memmove(void *dest, const void *src, uint64_t n) {
     uint8_t *pdest = (uint8_t *) dest;
     const uint8_t *psrc = (const uint8_t *) src;
 
     if (src > dest) {
-        for (size_t i = 0; i < n; i++) {
+        for (uint64_t i = 0; i < n; i++) {
             pdest[i] = psrc[i];
         }
     } else if (src < dest) {
-        for (size_t i = n; i > 0; i--) {
+        for (uint64_t i = n; i > 0; i--) {
             pdest[i - 1] = psrc[i - 1];
         }
     }
@@ -40,11 +40,11 @@ void *memmove(void *dest, const void *src, size_t n) {
     return dest;
 }
 
-int memcmp(const void *s1, const void *s2, size_t n) {
+int memcmp(const void *s1, const void *s2, uint64_t n) {
     const uint8_t *p1 = (const uint8_t *) s1;
     const uint8_t *p2 = (const uint8_t *) s2;
 
-    for (size_t i = 0; i < n; i++) {
+    for (uint64_t i = 0; i < n; i++) {
         if (p1[i] != p2[i]) {
             return p1[i] < p2[i] ? -1 : 1;
         }
@@ -52,8 +52,8 @@ int memcmp(const void *s1, const void *s2, size_t n) {
 
     return 0;
 }
-size_t strlen(const char *str) {
-    size_t length = 0;
+uint64_t strlen(const char *str) {
+    uint64_t length = 0;
 
     while (str[length] != '\0') {
         length++;
@@ -80,8 +80,8 @@ char *strcat(char *dest, const char *src) {
     return original_dest;
 }
 
-int strncmp(const char *s1, const char *s2, size_t n) {
-    for (size_t i = 0; i < n; i++) {
+int strncmp(const char *s1, const char *s2, uint64_t n) {
+    for (uint64_t i = 0; i < n; i++) {
         if (s1[i] != s2[i] || s1[i] == '\0') {
             return s1[i] < s2[i] ? -1 : 1;
         }
@@ -89,9 +89,9 @@ int strncmp(const char *s1, const char *s2, size_t n) {
     return 0;
 }
 
-char *strncpy(char *dest, const char *src, size_t n) {
+char *strncpy(char *dest, const char *src, uint64_t n) {
     char *original_dest = dest;
-    size_t i;
+    uint64_t i;
     for (i = 0; i < n && src[i] != '\0'; i++) {
         dest[i] = src[i];
     }
@@ -101,9 +101,9 @@ char *strncpy(char *dest, const char *src, size_t n) {
     return original_dest;
 }
 
-void *memchr(const void *s, int c, size_t n) {
+void *memchr(const void *s, int c, uint64_t n) {
     const uint8_t *p = (const uint8_t *) s;
-    for (size_t i = 0; i < n; i++) {
+    for (uint64_t i = 0; i < n; i++) {
         if (p[i] == (uint8_t) c) {
             return (void *) (p + i);
         }
@@ -111,9 +111,9 @@ void *memchr(const void *s, int c, size_t n) {
     return NULL;
 }
 
-void *memrchr(const void *s, int c, size_t n) {
+void *memrchr(const void *s, int c, uint64_t n) {
     const uint8_t *p = (const uint8_t *) s;
-    for (size_t i = n; i > 0; i--) {
+    for (uint64_t i = n; i > 0; i--) {
         if (p[i - 1] == (uint8_t) c) {
             return (void *) (p + (i - 1));
         }
